Split files and cleaned formatting
