+++ CMakeLists.txt
@@ -153,9 +153,13 @@ build_lib(
     model/value-axis.cc
     model/xy-series.cc
     model/throughput-sink.cc
-    model/net-visualizer.cc
-    helper/link-pools.cc
-  HEADER_FILES
+    model/series-wrapper.cc
+    model/series-wrapper-collection.cc
+    model/series-manager.cc
+    helper/logical-link-pairs.cc
+    helper/logical-link-paths.cc
+
+    HEADER_FILES
     helper/area-helper.h
     helper/building-configuration-container.h
     helper/building-configuration-helper.h
@@ -187,8 +191,12 @@ build_lib(
     model/value-axis.h
     model/xy-series.h
     model/throughput-sink.h
-    model/net-visualizer.h
-    helper/link-pools.h
+    model/series-wrapper.h
+    model/series-wrapper-collection.h
+    model/series-manager.h
+    helper/logical-link-pairs.h
+    helper/logical-link-paths.h
+
   LIBRARIES_TO_LINK
     ${libbuildings}
     ${libcore}
