+++ helper/logical-link-paths.h
@@ -31,8 +31,8 @@
  * Author: Andrew Wagger <andrew.wagger@nist.gov>
  */
 
-#ifndef LOGICAL_LINK_POOLS_H
-#define LOGICAL_LINK_POOLS_H
+#ifndef LOGICAL_LINK_PATHS_H
+#define LOGICAL_LINK_PATHS_H
 
 #include "../model/orchestrator.h"
 
@@ -94,9 +94,9 @@ class LogicalLinkPaths : public Object
                  const std::unordered_map<std::string, Ptr<AttributeValue>>& attributes);
 
   private:
-
     /**
-     * A vector of vectors containing the LogicalLinks for all paths. vectors may be longer than needed
+     * A vector of vectors containing the LogicalLinks for all paths. vectors may be longer than
+     * needed
      */
     std::vector<std::vector<Ptr<netsimulyzer::LogicalLink>>> m_paths;
 
@@ -106,139 +106,6 @@ class LogicalLinkPaths : public Object
     Ptr<Orchestrator> m_orchestrator;
 };
 
-class LogicalLinkPairs : public Object
-{
-  public:
-    LogicalLinkPairs(Ptr<Orchestrator> orchestrator, uint32_t nodeCount);
-    /**
-     * @brief Gets the L2 link between the two nodes if it exists
-     * @param i,j
-     * The nodes to check for a connection between. Identical if swapped
-     */
-    Ptr<netsimulyzer::LogicalLink> GetLink(uint32_t i, uint32_t j);
-
-    /**
-     * @brief Creates/Sets the color of a L2 Link
-     * @param i,j
-     * The Nodes to connect. Identical if swapped
-     */
-    void SetLink(uint32_t i, uint32_t j);
-
-    /**
-     * @brief Creates/Sets the color of a L2 Link
-     * @param i,j
-     * The Nodes to connect. Identical if swapped
-     * @param color
-     * The color to make the link
-     */
-    void SetLink(uint32_t i, uint32_t j, netsimulyzer::Color3 color);
-
-    /**
-     * @brief Creates/Sets the color of a L2 Link
-     * @param i,j
-     * The Nodes to connect. Identical if swapped
-     * @param color
-     * The color to make the link
-     * @param attributes
-     * Attributes to be given to this link
-     */
-    void SetLink(uint32_t i,
-                 uint32_t j,
-                 netsimulyzer::Color3 color,
-                 const std::unordered_map<std::string, Ptr<AttributeValue>>& attributes);
-
-    /**
-     * @brief If a L2 link exists between the two nodes, deactivate it
-     * @param i,j
-     * The nodes to check for a connection between. Identical if swapped
-     */
-    void RemoveLink(uint32_t i, uint32_t j);
-
-  private:
-    template <typename T>
-    class PairMap
-    {
-      private:
-        uint32_t m_n;
-        uint32_t m_size;
-        T* m_arr;
-
-      public:
-        PairMap(void)
-            : m_n(0),
-              m_size(0),
-              m_arr(nullptr) {};
-
-        PairMap(uint32_t n)
-            : m_n(n)
-        {
-            this->m_size = n * (n - 1) / 2;
-            this->m_arr = new T[this->m_size];
-        };
-
-        ~PairMap()
-        {
-            delete[] this->m_arr;
-        };
-
-        T Get(uint32_t i, uint32_t j)
-        {
-            if (i > j)
-            {
-                return this->Get(j, i);
-            }
-            else if (i >= this->m_n || j >= this->m_n || i == j)
-            {
-                throw "oob";
-            }
-            else
-            {
-                uint32_t index =
-                    (j - (i + 1)) + (this->Length() - ((this->m_n - i) * (this->m_n - i - 1) / 2));
-                return this->m_arr[index];
-            }
-        };
-
-        bool Set(uint32_t i, uint32_t j, T value)
-        {
-            if (i > j)
-            {
-                return this->Set(j, i, value);
-            }
-            else if (i >= this->m_n || j >= this->m_n || i == j)
-            {
-                return false;
-            }
-            else
-            {
-                this->m_arr[(j - (i + 1)) +
-                            (this->Length() - ((this->m_n - i) * (this->m_n - i - 1) / 2))] = value;
-                return false;
-            }
-        };
-
-        uint32_t Length()
-        {
-            return this->m_size;
-        };
-
-        uint32_t N()
-        {
-            return this->m_n;
-        };
-    };
-
-    /**
-     * A data structure that stores the pairwise LogicalLinks
-     */
-    PairMap<Ptr<netsimulyzer::LogicalLink>> m_pairMap;
-
-    /**
-     * Orchestrator attatched to the LogicalLinks
-     */
-    Ptr<Orchestrator> m_orchestrator;
-};
-
 } // namespace ns3::netsimulyzer
 
-#endif // LOGICAL_LINK_POOLS_H
+#endif // LOGICAL_LINK_PATHS_H
