+++ model/series-wrapper.cc
@@ -31,7 +31,7 @@
  * Author: Andrew Wagger <andrew.wagger@nist.gov>
  */
 
-#include "net-visualizer.h"
+#include "series-wrapper.h"
 
 #include "color.h"
 #include "netsimulyzer-3D-models.h"
@@ -58,7 +58,7 @@
 namespace ns3
 {
 
-NS_LOG_COMPONENT_DEFINE("Visualizer");
+NS_LOG_COMPONENT_DEFINE("SeriesWrapper");
 
 namespace netsimulyzer
 {
@@ -68,9 +68,6 @@ NS_OBJECT_ENSURE_REGISTERED(AccumulatorWrapper);
 NS_OBJECT_ENSURE_REGISTERED(AverageValueWrapper);
 NS_OBJECT_ENSURE_REGISTERED(SlidingValueWrapper);
 NS_OBJECT_ENSURE_REGISTERED(SlidingLoadWrapper);
-NS_OBJECT_ENSURE_REGISTERED(SeriesWrapperCollection);
-NS_OBJECT_ENSURE_REGISTERED(SeriesWrapperMap);
-NS_OBJECT_ENSURE_REGISTERED(SeriesManager);
 
 /*          SeriesWrapper           */
 
@@ -107,214 +104,6 @@ SeriesWrapper::operator()()
 {
     return this->GetSeries();
 };
-
-/*          SeriesWrapperCollection           */
-
-TypeId
-SeriesWrapperCollection::GetTypeId()
-{
-    // clang-format off
-    static TypeId tid =
-        TypeId ("ns3::netsimulyzer::SeriesWrapperCollection")
-            .SetParent<ns3::Object> ()
-            .SetGroupName ("netsimulyzer");
-    return tid;
-    // clang-format on
-}
-
-SeriesWrapperCollection*
-SeriesWrapperCollection::AddWrapper(Ptr<SeriesWrapper> w)
-{
-    NS_LOG_FUNCTION(this << w);
-    m_wrappers.emplace_back(w);
-    m_collection->Add(w->GetSeries());
-    return this;
-};
-
-Ptr<XYSeries>
-SeriesWrapperCollection::GetSeries(std::size_t i)
-{
-    NS_LOG_FUNCTION(this << i);
-    return m_wrappers.at(i)->GetSeries();
-};
-
-Ptr<SeriesWrapper>
-SeriesWrapperCollection::GetWrapper(std::size_t i)
-{
-    NS_LOG_FUNCTION(this << i);
-    return m_wrappers.at(i);
-};
-
-std::size_t
-SeriesWrapperCollection::GetNSeries()
-{
-    return m_wrappers.size();
-};
-
-Ptr<SeriesCollection>
-SeriesWrapperCollection::GetSeriesCollection()
-{
-    return m_collection;
-};
-
-SeriesWrapperCollection::SeriesWrapperCollection(Ptr<SeriesManager> manager)
-{
-    NS_LOG_FUNCTION(this << manager);
-    m_collection = CreateObject<SeriesCollection>(manager->GetOrchestrator());
-};
-
-SeriesWrapperCollection::SeriesWrapperCollection(Ptr<SeriesManager> manager,
-                                 std::string name,
-                                 std::string x_axis,
-                                 std::string y_axis)
-    : SeriesWrapperCollection(manager)
-{
-    NS_LOG_FUNCTION(this << manager << name << x_axis << y_axis);
-
-    m_collection->SetAttribute("Name", StringValue(name));
-    m_collection->GetYAxis()->SetAttribute("Name", StringValue(y_axis));
-    m_collection->GetXAxis()->SetAttribute("Name", StringValue(x_axis));
-};
-
-std::vector<Ptr<SeriesWrapper>>::iterator
-SeriesWrapperCollection::begin()
-{
-    return m_wrappers.begin();
-};
-
-std::vector<Ptr<SeriesWrapper>>::iterator
-SeriesWrapperCollection::end()
-{
-    return m_wrappers.end();
-};
-
-SeriesWrapper&
-SeriesWrapperCollection::operator[](std::size_t i)
-{
-    return (*this->GetWrapper(i));
-};
-
-/*          SeriesManager           */
-
-TypeId
-SeriesManager::GetTypeId()
-{
-    // clang-format off
-    static TypeId tid =
-        TypeId ("ns3::netsimulyzer::SeriesManager")
-            .SetParent<ns3::Object> ()
-            .SetGroupName ("netsimulyzer")
-            .AddAttribute("Orchestrator",
-                          "The Orchestrator for the SeriesManager",
-                          PointerValue(),
-                          MakePointerAccessor(&SeriesManager::GetOrchestrator),
-                          MakePointerChecker<Orchestrator>());
-    return tid;
-    // clang-format on
-}
-
-SeriesManager::SeriesManager(std::string outputFileName)
-    : m_orchestrator(CreateObject<Orchestrator>(outputFileName)),
-      m_configHelper(m_orchestrator)
-{
-    NS_LOG_FUNCTION(this << outputFileName);
-};
-
-SeriesManager::SeriesManager(std::string outputFileName, NodeContainer nodes)
-    : SeriesManager(outputFileName)
-{
-    NS_LOG_FUNCTION(this << outputFileName << &nodes);
-    SetNodes(nodes);
-};
-
-void
-SeriesManager::SetNodes(NodeContainer nodes)
-{
-    NS_LOG_FUNCTION(this << &nodes);
-    m_nodes = nodes;
-    m_configContainer = m_configHelper.Install(m_nodes);
-};
-
-Ptr<SeriesWrapperCollection>
-SeriesManager::GetCollection(std::string index)
-{
-    NS_LOG_FUNCTION(this << index);
-    return m_collections.at(index);
-};
-
-SeriesWrapperCollection&
-SeriesManager::operator[](std::string index)
-{
-    return (*GetCollection(index));
-};
-
-SeriesManager*
-SeriesManager::SetContainer(std::string index, Ptr<SeriesWrapperCollection> container)
-{
-    NS_LOG_FUNCTION(this << index << container);
-    m_collections.insert({index, container});
-    return this;
-};
-
-Ptr<NodeConfiguration>
-SeriesManager::GetConfig(std::size_t i)
-{
-    NS_LOG_FUNCTION(this << i);
-    return m_configContainer.Get(i);
-};
-
-NodeConfigurationHelper*
-SeriesManager::ConfigHelper()
-{
-    return &m_configHelper;
-}
-
-
-Ptr<Node>
-SeriesManager::GetNode(std::size_t i)
-{
-    NS_LOG_FUNCTION(this << i);
-    return m_nodes.Get(i);
-};
-
-Ptr<Node>
-SeriesManager::GetNodeById(uint32_t id)
-{
-    NS_LOG_FUNCTION(this << id);
-    for (std::size_t i = 0; i < m_nodes.GetN(); ++i)
-    {
-        if (m_nodes.Get(i)->GetId() == id)
-        {
-            return m_nodes.Get(i);
-        }
-    }
-    return nullptr;
-}
-
-std::size_t
-SeriesManager::GetNNodes()
-{
-    return m_nodes.GetN();
-};
-
-Ptr<Orchestrator>
-SeriesManager::GetOrchestrator() const
-{
-    return m_orchestrator;
-};
-
-std::unordered_map<std::string, Ptr<SeriesWrapperCollection>>::iterator
-SeriesManager::begin()
-{
-    return m_collections.begin();
-};
-
-std::unordered_map<std::string, Ptr<SeriesWrapperCollection>>::iterator
-SeriesManager::end()
-{
-    return m_collections.end();
-};
-
 
 /*          AccumulatorWrapper           */
 
@@ -437,7 +226,9 @@ SlidingValueWrapper::SlidingValueWrapper(Ptr<XYSeries> series, Time interval)
     }
 };
 
-SlidingValueWrapper::SlidingValueWrapper(Ptr<XYSeries> series, Time window, double maxSampleFrequency)
+SlidingValueWrapper::SlidingValueWrapper(Ptr<XYSeries> series,
+                                         Time window,
+                                         double maxSampleFrequency)
     : SlidingValueWrapper(series)
 {
     NS_LOG_FUNCTION(this << series << maxSampleFrequency);
@@ -465,7 +256,7 @@ SlidingValueWrapper::Update(Time time, double value)
             m_values.erase(pair--);
         }
     }
-    
+
     m_values.push_back({time, value});
     if (!m_timed)
     {
@@ -539,16 +330,19 @@ SlidingLoadWrapper::SlidingLoadWrapper(Ptr<XYSeries> series, Time interval)
 };
 
 SlidingLoadWrapper::SlidingLoadWrapper(Ptr<XYSeries> series,
-                         Time window,
-                         double bandwidth,
-                         double maxSampleFrequency)
+                                       Time window,
+                                       double bandwidth,
+                                       double maxSampleFrequency)
     : SlidingValueWrapper(series, window, maxSampleFrequency),
       m_bandwidth(bandwidth)
 {
     NS_LOG_FUNCTION(this << series << bandwidth);
 };
 
-SlidingLoadWrapper::SlidingLoadWrapper(Ptr<XYSeries> series, Time window, double bandwidth, Time interval)
+SlidingLoadWrapper::SlidingLoadWrapper(Ptr<XYSeries> series,
+                                       Time window,
+                                       double bandwidth,
+                                       Time interval)
     : SlidingValueWrapper(series, window, interval),
       m_bandwidth(bandwidth)
 {
@@ -572,64 +366,6 @@ SlidingLoadWrapper::GetSlidingValue() const
 {
     return SlidingValueWrapper::GetSlidingValue();
 }
-
-/*          SeriesWrapperMap           */
-
-TypeId
-SeriesWrapperMap::GetTypeId()
-{
-    // clang-format off
-    static TypeId tid =
-        TypeId ("ns3::netsimulyzer::SeriesWrapperMap")
-            .SetParent<ns3::netsimulyzer::SeriesWrapperCollection> ()
-            .SetGroupName ("netsimulyzer");
-    return tid;
-    // clang-format on
-}
-
-SeriesWrapperMap::SeriesWrapperMap(Ptr<SeriesManager> manager)
-    : SeriesWrapperCollection(manager)
-{
-    NS_LOG_FUNCTION(this << manager);
-};
-
-SeriesWrapperMap::SeriesWrapperMap(Ptr<SeriesManager> manager,
-                     std::string name,
-                     std::string x_axis,
-                     std::string y_axis)
-    : SeriesWrapperCollection(manager, name, x_axis, y_axis)
-{
-    NS_LOG_FUNCTION(this << manager << name << x_axis << y_axis);
-};
-
-SeriesWrapperMap*
-SeriesWrapperMap::AddWrapper(std::string index, Ptr<SeriesWrapper> w)
-{
-    NS_LOG_FUNCTION(this << index << w);
-    m_nameMap.insert({index, GetNSeries()});
-    SeriesWrapperCollection::AddWrapper(w);
-    return this;
-};
-
-Ptr<XYSeries>
-SeriesWrapperMap::GetSeries(std::string index)
-{
-    NS_LOG_FUNCTION(this << index);
-    return SeriesWrapperCollection::GetSeries(m_nameMap.at(index));
-};
-
-Ptr<SeriesWrapper>
-SeriesWrapperMap::GetWrapper(const std::string& index)
-{
-    NS_LOG_FUNCTION(this << index);
-    return SeriesWrapperCollection::GetWrapper(m_nameMap.at(index));
-};
-
-SeriesWrapper&
-SeriesWrapperMap::operator[](const std::string& index)
-{
-    return (*this->GetWrapper(index));
-};
 
 } // namespace netsimulyzer
